Support opening non existant files using untitled:// scheme
Rely on untitled scheme from https://github.com/eclipse-theia/theia/pull/10868
to create an editor for a file that will be saved as the path passed on
the command line.

The protocol was changed to use `uri` instead of `filePath`. The python
client was updated to keep compatibility with the new protocol, but the
feature of opening non existent files is only implemented in the nodejs
client.